refactor: [US-057] - Move DNA tasks into core/tasks/dna.py (wire names pinned)
