+++ scripts/ralph/prd.json
@@ -958,7 +958,7 @@
         "**Celery autodiscovery check:** in a dev container, run `celery -A bibliotype inspect registered` (or, if a worker isn't running, `python -c \"from bibliotype.celery import app; print([t for t in app.tasks if t.startswith('core.')])\"`) and confirm both `core.tasks.generate_dna_task` and `core.tasks.claim_anonymous_dna_task` appear. Capture the output line in `progress.txt`.",
         "`manage.py test core.tests.test_tasks_integration` passes."
       ],
-      "passes": false,
+      "passes": true,
       "notes": "",
       "priority": 66
     },
